+++ odoo/custom/src/private/faf_sale/__manifest__.py
@@ -10,6 +10,7 @@
     "author": "Tharathip C., Ecosoft",
     "depends": ["sale"],
     "data": [
+        "views/res_config_settings_views.xml",
         "views/sale_views.xml",
     ],
     "license": "AGPL-3",